refactor: extract skills panel from AgentWorkspace
